requirements.txt
alembic==1.6.5
asgiref==3.5.2
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.1
distlib==0.3.4
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
dnspython==2.1.0
email-validator==1.1.3
filelock==3.6.0
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
Pillow==9.2.0
pipenv==2022.4.30
platformdirs==2.5.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.23
sqlparse==0.4.2
tk==0.1.0
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
Werkzeug==2.0.1
WTForms==2.3.3
